Move activities to a special package
