+++ dontforget-android/AndroidManifest.xml
@@ -16,15 +16,15 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".ConnectionActivity"
+            android:name=".activity.ConnectionActivity"
             android:label="@string/app_name" >
         </activity>
         <activity
-            android:name=".TasksActivity"
+            android:name=".activity.TasksActivity"
             android:label="@string/title_activity_tasks" >
         </activity>
         <activity
-            android:name=".WelcomeActivity"
+            android:name=".activity.WelcomeActivity"
             android:label="@string/title_activity_welcome" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
