+++ dontforget-android/res/layout/activity_connection.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="fr.chklang.dontforget.android.ConnectionActivity" >
+    tools:context="fr.chklang.dontforget.android.activity.ConnectionActivity" >
 
     <Spinner
         android:id="@+id/connection_protocol"
