+++ dontforget-android/res/layout/activity_welcome.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="fr.chklang.dontforget.android.WelcomeActivity" >
+    tools:context="fr.chklang.dontforget.android.activity.WelcomeActivity" >
 
     <TextView
         android:layout_width="wrap_content"
